Separate series and data point components
